requirements.txt
datasets
dvc[azure]
elasticsearch-dsl
faiss-cpu
fuzzywuzzy
git+https://github.com/srush/MiniChain/@v0.1
GitPython
manifest-ml
minichain
numpy
openai
pandas
Pillow
pymupdf
streamlit
tiktoken
guardrails-ai
pydantic
